+++ src/components/TodoLayout/TodoList/components/TodoListHeader/TodoListHeader.tsx
@@ -1,19 +1,22 @@
 import { NotificationType } from "@/common/constants/NotificationConstants.ts";
-import { CurrentTodoFilter, TodoErrorMessages } from "@/common/constants/TodoConstants.ts";
-import { FilteredTodo } from "@/components/TodoLayout/TodoList/components/FilteredTodo/FilteredTodo.tsx";
+import { TodoCurrentFilter } from "@/common/constants/TodoConstants/TodoFilters.ts";
+import { TodoConfirmMessages, TodoNotificationMessages } from "@/common/constants/TodoConstants/TodoValidation.ts";
+import { TodoFiltered } from "@/components/TodoLayout/TodoList/components/TodoFiltered/TodoFiltered.tsx";
 import { Button } from "@/components/ui/Button/Button.tsx";
 import { useAppDispatch } from "@/hooks/useAppDispatch.ts";
+import { useModalState } from "@/hooks/useModalState.ts";
 import { setNotification } from "@/store/actions/notificationActionCreators.ts";
 import { deleteCompletedTodos, setFiltrationValue } from "@/store/actions/todoActionCreators.ts";
 import cn from "classnames";
 import styles from "./TodoListHeader.module.scss";
 import { useAppSelector } from "@/hooks/useAppSelector.ts";
 import { useEffect, useState } from "react";
-import filteredStyles from "@/components/TodoLayout/TodoList/components/FilteredTodo/FilteredTodo.module.scss";
+import filteredStyles from "@/components/TodoLayout/TodoList/components/TodoFiltered/TodoFiltered.module.scss";
 
 export const TodoListHeader = () => {
   const { originalTodos } = useAppSelector(state => state.todoReducer);
   const [completedTodo, setCompletedTodo] = useState(0);
+  const setConfirmModalActive = useModalState("confirmModal")[1];
   const dispatch = useAppDispatch();
 
   useEffect(() => {
@@ -22,9 +25,14 @@ export const TodoListHeader = () => {
   }, [originalTodos]);
 
   const handleDeleteCompletedTodos = () => {
-    dispatch(setNotification({ title: TodoErrorMessages.CLEAR_COMPLETED, type: NotificationType.SUCCESS }));
-    dispatch(deleteCompletedTodos());
-    dispatch(setFiltrationValue(CurrentTodoFilter.ALL));
+    setConfirmModalActive(true, {
+      confirmCallback: () => {
+        dispatch(setNotification({ title: TodoNotificationMessages.DELETE_COMPLETED_TODOS, type: NotificationType.SUCCESS }));
+        dispatch(deleteCompletedTodos());
+        dispatch(setFiltrationValue(TodoCurrentFilter.ALL));
+      },
+      message: TodoConfirmMessages.DELETE_COMPLETED_TODOS,
+    });
   };
 
   return (
@@ -50,7 +58,7 @@ export const TodoListHeader = () => {
           Clear Completed
         </Button>
       </div>
-      <FilteredTodo />
+      <TodoFiltered />
     </div>
   );
 };
